backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-core
pypdf
python-docx
beautifulsoup4
requests
python-dotenv
openai
